// src/layout/StickyGrid/interfaces.ts
export interface IThumbnail {
  className?: string;
  style?: React.CSSProperties;
  id: string;
  placeholderLabel?: string;
  placeholderLabelStyle?: React.CSSProperties;
  placeholderLabelRequired?: boolean;
  placeholderIcon?: string;
  placeholderIconStyle?: React.CSSProperties;
  previewStyle?: React.CSSProperties;
  previewClassName?: string;
  previewMimeType: string;
  previewSrcUrl: string;
  previewMouseSensitive?: boolean;
  children?: JSX.Element | React.ReactNode;
  overlay?: {
    icon?: string;
    text?: string;
    tooltip?: string | JSX.Element | string[];
  };
  contextmenu?: {
    title?: string;
    items?: IPopoverListItem[];
    originAnchor?: PopoverOrigin;
    originTransf?: PopoverOrigin;
  };
}

export interface ICellClick {
  eventDetail: number;
  columnIndex: number;
  rowIndex: number;
  selected: boolean;
  highligh: boolean;
  keyDownCtrl: boolean;
  keyDownMeta: boolean;
  thumbId?: string;
}